Added Django webframework runing under twisted.  Added debug flag and dmsg() function for debugging. Cleaned up hardcoded URLs.
Did a minimum amount of work to make sure I could deploy with Django to the "production" rasberry pi.  This required creating
a local_settings.py for hardcoded paths and database names.   Got it all working under Twisted instead of Django using a separate
runserver process.

Because I notice images are slow to advance on the physical hardware of the raspberry pi, I added a global DEBUG flag, a keyboard handler to toggle it, and
a wrapper around msg() called dmsg() which only shows a message if DEBUG > 1.