Added setting FRAMEIDENTITY and display it in the template.
This feature is to help tell one frame apart from another merely by
glancing at it's remote.  The FRAMEIDENTITY is an arbitrary string
and can be set in the settings file or overridden in the local settings file.

The expected use is to configure this setting in the local settings file.